kochapp/requirements.txt
djoser==2.1.0
django==3.1.3
django-filter==2.3.0
djangorestframework==3.11.1
Pillow==7.2.0
psycopg2==2.8.6
python-decouple==3.3
drf-yasg==1.17.1
gunicorn==20.0.4
django-cors-headers==3.5.0
